chAIagent/requirements.txt
agno
ollama
uvicorn
clickhouse-driver
aiosqlite
fastapi
faiss-cpu 
sentence-transformers
